refactor(pallets/communities): split types into multiple files
